Replace node 6 with node 11 in travis
npm 3 used by node 6 does not automatically run the prepare script
specified in the package.json

Because of that we can't rely on it to be executed and therefore our
tests will fail. We've fixed it for circleci which is enough to cover
support for it.

To provide even more coverage, we'll also test on node 11 instead of
node 6 in travis.