Ensure ~ operator converts to string
Ensure that the binary concatenation operator (~) always produces a string value.
We do that by utilizing the string template syntax instead of the binary plus (+) operator
in the transpiled code.

To make the generated code more efficient and easier to read we'll also collapse
sibling template literals into a single.

fixes #123